// src/server/api/models/banking/models/BillSchedule.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { BillPayment } from './BillPayment';
import type { BillSchedulePaymentFrequency } from './BillSchedulePaymentFrequency';
import type { BillScheduleStatus } from './BillScheduleStatus';
import type { LocalDate } from './LocalDate';
import type { LocalDateTime } from './LocalDateTime';
import type { SubAccountType } from './SubAccountType';
import type { UUID } from './UUID';

export type BillSchedule = {
    /**
     * The unique ID for the record. This is the ID that is used in all services and related records when referencing the record.
     */
    id: UUID;
    /**
     * The ID of the user that created the record.
     */
    createdBy: string;
    /**
     * When the record was created.
     */
    createdAt: LocalDateTime;
    /**
     * The ID of the user that last modified the record.
     */
    updatedBy: string;
    /**
     * When the record was last modified.
     */
    updatedAt: LocalDateTime;
    /**
     * Indicates if this is a deleted record
     */
    deleted: boolean;
    /**
     * The client ID from the crm service
     */
    clientId: UUID;
    /**
     * The client description
     */
    clientLabel: string;
    /**
     * The client code from CRM
     */
    clientCode: string;
    payments: Array<BillPayment>;
    /**
     * A unique schedule code generated by the system
     */
    code: string;
    /**
     * The bill amount
     */
    billAmount: number;
    /**
     * The bill currency
     */
    currency: string;
    /**
     * The bill description
     */
    label: string;
    /**
     * The status of the schedule
     */
    status: BillScheduleStatus;
    /**
     * The schedule frequency
     */
    paymentFrequency: BillSchedulePaymentFrequency;
    /**
     * The start date
     */
    startDate: LocalDate;
    /**
     * The end date. This is generated by adding the payment frequency to the start date.
     */
    endDate: LocalDate;
    /**
     * The idempotency key
     */
    idemPotencyId?: UUID | null;
    /**
     * The flag indicating if the request was auto approved
     */
    autoApprove: boolean;
    /**
     * The sub account type
     */
    subAccountType: SubAccountType;
    /**
     * The deposit or loan account ID used for processing the scheduled payment
     */
    accountId: UUID;
    /**
     * The deposit or loan account label used for processing the scheduled payment
     */
    accountLabel: string;
    /**
     * The deposit or loan number used for processing the scheduled payment
     */
    accountNo: string;
    /**
     * The ID of the biller to be paid
     */
    billerId: UUID;
    /**
     * The code of the biller to be paid
     */
    billerCode: string;
    /**
     * The label of the biller to be paid
     */
    billerLabel: string;
};